+++ packages/@cldcvr/flow-form-builder/src/components/f-form-builder/fields/textarea.ts
@@ -24,6 +24,7 @@ export default function (
 			?clear=${field?.clear ?? true}
 			?read-only=${field?.readonly ?? false}
 			?resizable=${field?.resizable ?? false}
+			?mask-value=${field?.maskValue ?? false}
 			rows=${ifDefined(field?.rows)}
 			@click=${ifDefined(field.onClick)}
 			@focus=${ifDefined(field.onFocus)}